refactoring(fare): step2 FarePolicy 패키지 변경 및 상수로 변경
