--- src/main/java/nextstep/subway/maps/fare/domain/FarePolicy.java
@@ -1,5 +1,0 @@
-package nextstep.subway.maps.fare.domain;
-
-public interface FarePolicy {
-    void calculate(FareContext fareContext);
-}
