+++ src/main/java/nextstep/subway/maps/fare/domain/policy/DiscountByAgeFarePolicy.java
@@ -1,8 +1,14 @@
-package nextstep.subway.maps.fare.domain;
+package nextstep.subway.maps.fare.domain.policy;
 
+import nextstep.subway.maps.fare.domain.Fare;
+import nextstep.subway.maps.fare.domain.FareContext;
 import nextstep.subway.members.member.domain.Member;
 
 public class DiscountByAgeFarePolicy implements FarePolicy {
+    private static final int DISCOUNT_AGE_MAX = 19;
+    private static final int DISCOUNT_YOUTH_AGE_MIN = 13;
+    private static final int DISCOUNT_CHILDREN_AGE_MIN = 6;
+
     @Override
     public void calculate(FareContext fareContext) {
         Member member = fareContext.getMember();
@@ -13,18 +19,20 @@ public class DiscountByAgeFarePolicy implements FarePolicy {
         Integer age = member.getAge();
         Fare fare = fareContext.getFare();
 
-        if (age >= 19) {
+        if (age >= DISCOUNT_AGE_MAX) {
             return;
         }
 
-        if (age >= 13) {
+        if (age >= DISCOUNT_YOUTH_AGE_MIN) {
             fare.discountFare(350);
             fare.discountPercent(20);
+            return;
         }
 
-        if (age >= 6 && age < 13) {
+        if (age >= DISCOUNT_CHILDREN_AGE_MIN) {
             fare.discountFare(350);
             fare.discountPercent(50);
+            return;
         }
 
         fare.discountPercent(100);
