+++ src/main/java/nextstep/subway/maps/fare/domain/policy/FarePolicy.java
@@ -1,0 +1,7 @@
+package nextstep.subway.maps.fare.domain.policy;
+
+import nextstep.subway.maps.fare.domain.FareContext;
+
+public interface FarePolicy {
+    void calculate(FareContext fareContext);
+}
